separate out chatlistitem and add muted icon
